fix: bfloat16

+++ deepseek_math500_eval.py
@@ -16,7 +16,7 @@ model = AutoModelForCausalLM.from_pretrained(
     "deepseek-ai/DeepSeek-R1", 
     trust_remote_code=True, 
     dir="./.cache/models",
-    torch_dtype=torch.float16  # Use FP16 instead of FP8
+    torch_dtype=torch.bfloat16  # Use FP16 instead of FP8
 )
 
 # model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, torch_dtype=torch.float16, device_map="auto", dir="./.cache/models")
